Added barrelling for duration
